requirements/base.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements/base.in
#
adrf==0.1.9
    # via -r requirements/base.in
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   mcp
    #   openai
    #   sse-starlette
    #   starlette
asgiref==3.8.1
    # via django
async-property==0.2.2
    # via adrf
certifi==2025.4.26
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via uvicorn
colorama==0.4.6
    # via griffe
distro==1.9.0
    # via openai
django==5.2
    # via
    #   -r requirements/base.in
    #   adrf
    #   djangorestframework
django-environ==0.12.0
    # via -r requirements/base.in
djangorestframework==3.16.0
    # via adrf
griffe==1.7.3
    # via openai-agents
gunicorn==23.0.0
    # via
    #   -r requirements/base.in
    #   uvicorn-worker
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.6.4
    # via -r requirements/base.in
httpx==0.28.1
    # via
    #   mcp
    #   openai
httpx-sse==0.4.0
    # via mcp
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
jiter==0.9.0
    # via openai
mcp==1.6.0
    # via openai-agents
openai==1.76.2
    # via
    #   -r requirements/base.in
    #   openai-agents
openai-agents==0.0.14
    # via -r requirements/base.in
packaging==25.0
    # via gunicorn
psycopg2-binary==2.9.10
    # via -r requirements/base.in
pydantic==2.11.4
    # via
    #   mcp
    #   openai
    #   openai-agents
    #   pydantic-settings
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.9.1
    # via mcp
python-dotenv==1.1.0
    # via pydantic-settings
requests==2.32.3
    # via openai-agents
sniffio==1.3.1
    # via
    #   anyio
    #   openai
sqlparse==0.5.3
    # via django
sse-starlette==2.3.3
    # via mcp
starlette==0.46.2
    # via
    #   mcp
    #   sse-starlette
structlog==25.3.0
    # via -r requirements/base.in
tqdm==4.67.1
    # via openai
types-requests==2.32.0.20250328
    # via openai-agents
typing-extensions==4.13.2
    # via
    #   openai
    #   openai-agents
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.0
    # via
    #   pydantic
    #   pydantic-settings
urllib3==2.4.0
    # via
    #   requests
    #   types-requests
uvicorn==0.34.2
    # via
    #   -r requirements/base.in
    #   mcp
    #   uvicorn-worker
uvicorn-worker==0.3.0
    # via -r requirements/base.in
uvloop==0.21.0
    # via -r requirements/base.in
